refactor: rename board, task models files
